Implement Phase 1: Consolidate Config into Daemon Package
- Move YAML config types (File, Logging, QueueStorage, etc.) to daemon/config.go
- Move ApplyConfigFile() and helper functions to daemon package
- Update cmd/querator/server.go to use daemon.File and daemon.ApplyConfigFile
- Move config tests to daemon/config_test.go
- Delete config/ package entirely
